# perf(Role): linear speed position getter