**Please describe the changes this PR makes and why it should be merged:**

Alternative to #9490, fixes #3925.

| Iterations |      Old |      New | Difference |
| :--------: | :------: | :------: | :--------: |
|     10_000 | 631.24ms | 174.44ms |     27.63% |
|    100_000 |    8.01s |    1.89s |     23.59% |

Basically skips the `sorted` call, which is an algorithm with an O(n * log(n)) cost, as well as the indexOf, which is a O(n) call. With this, not only it's not longer allocating a Collection nor an Array (making the only allocation be the iterator used in reduce), but the performance also goes down to a single O(n).

<!--
copilot:poem
-->
### <samp>🤖 Generated by Copilot at 8b28bc4</samp>

> _`position` getter_
> _improved with reduce function_
> _snowflakes in winter_

**Status and versioning classification:**

- Code changes have been tested against the Discord API, or there are no code changes
- I know how to update typings and have done so, or typings don't need updating