@kyranet 2023-05-01
Yes, in a follow-up PR, I wanted to push this one first to get reviews in case the code can be improved any further.