@jaw0r3k 2023-05-01
> Yes, in a follow-up PR, I wanted to push this one first to get reviews in case the code can be improved any further.

Idk if using simply for loop will be a little bit faster, as not calling the function everytime
Except that it looks good